training/pyproject.toml
[project]
name = "gemma3-training"
version = "0.1.0"
description = "Trigger train jobs and monitor status"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-storage>=2.10.0",
    "google-cloud-firestore>=2.21.0",
    "google-cloud-run>=0.11.0",
    "firebase-admin>=7.1.0",
    "setuptools",
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]